move update tests next to the implementation
